CodeEditor: Add codeeditor shortcode
We will support codemirror5 (cm5), codemirror6 (cm6) and ace for users
to choose the one they like.

Currently, only cm5 is available; cm6 lacks plenty of modes for
common languages, and has to be bundled by ourselves, but it's better
than cm5, we will see how it goes in the future; ace will be added
soon.